Piqa 82: risk analysis version from endpoint (#102)
