+++ features/catalog/step_definitions/e-service-risk-analysis-addition.ts
@@ -4,7 +4,6 @@ import { dataPreparationService } from "../../../services/data-preparation.servi
 import {
   assertContextSchema,
   getAuthorizationHeader,
-  getRiskAnalysis,
   getToken,
 } from "../../../utils/commons";
 import { EServiceMode } from "../../../api/models";
@@ -32,7 +31,10 @@ When("l'utente aggiunge un'analisi del rischio", async function () {
   });
   this.response = await apiClient.eservices.addRiskAnalysisToEService(
     this.eserviceId,
-    getRiskAnalysis({ completed: true, tenantType: this.tenantType }),
+    await dataPreparationService.getRiskAnalysis({
+      completed: true,
+      tenantType: this.tenantType,
+    }),
     getAuthorizationHeader(this.token)
   );
 });
@@ -54,7 +56,10 @@ When(
 
     this.response = await apiClient.eservices.addRiskAnalysisToEService(
       this.eserviceId,
-      getRiskAnalysis({ completed: true, tenantType }),
+      await dataPreparationService.getRiskAnalysis({
+        completed: true,
+        tenantType,
+      }),
       getAuthorizationHeader(this.token)
     );
   }
@@ -69,12 +74,13 @@ When(
       tenantType: TenantType,
     });
 
-    const { name, riskAnalysisForm } = getRiskAnalysis({
-      completed: true,
-      tenantType: this.tenantType,
-    });
+    const { name, riskAnalysisForm } =
+      await dataPreparationService.getRiskAnalysis({
+        completed: true,
+        tenantType: this.tenantType,
+      });
 
-    const outdatedVersion = "0.0";
+    const outdatedVersion = (Number(riskAnalysisForm.version) - 1).toFixed(1);
 
     this.response = await apiClient.eservices.addRiskAnalysisToEService(
       this.eserviceId,
