+++ features/purpose/step_definitions/purpose-common-steps.ts
@@ -10,7 +10,6 @@ import {
   assertContextSchema,
   getToken,
   getOrganizationId,
-  getRiskAnalysis,
   getRandomInt,
 } from "../../../utils/commons";
 import { TenantType } from "../../common-steps";
@@ -28,7 +27,7 @@ Given(
 
     const token = await getToken(tenantType);
     const consumerId = getOrganizationId(tenantType);
-    const { riskAnalysisForm } = getRiskAnalysis({
+    const { riskAnalysisForm } = await dataPreparationService.getRiskAnalysis({
       completed: true,
       tenantType,
     });
@@ -122,7 +121,10 @@ Given(
       await dataPreparationService.addRiskAnalysisToEService(
         token,
         this.eserviceId,
-        getRiskAnalysis({ completed: true, tenantType })
+        await dataPreparationService.getRiskAnalysis({
+          completed: true,
+          tenantType,
+        })
       );
   }
 );
@@ -157,7 +159,7 @@ Given(
     const token = await getToken(tenantType);
     const consumerId = getOrganizationId(tenantType);
 
-    const { riskAnalysisForm } = getRiskAnalysis({
+    const { riskAnalysisForm } = await dataPreparationService.getRiskAnalysis({
       completed: true,
       tenantType,
     });
