+++ utils/commons.ts
@@ -1,90 +1,9 @@
 import "../configs/env";
 import { readFileSync } from "fs";
 import { z } from "zod";
-import { AxiosResponse } from "axios";
-import { CreatedResource } from "../api/models";
-import { TenantType, Role, SessionTokens } from "../features/common-steps";
-import { apiClient } from "../api";
+import { type AxiosResponse } from "axios";
+import { TenantType, SessionTokens, Role } from "../features/common-steps";
 import { generateSessionTokens } from "./session-tokens";
-
-type RiskAnalysisTemplateType = "PA" | "Privato/GSP";
-
-const RISK_ANALYSIS_DATA: Record<
-  RiskAnalysisTemplateType,
-  { version: string; completed: unknown; uncompleted: unknown }
-> = {
-  "Privato/GSP": {
-    version: "2.0",
-    completed: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      usesPersonalData: ["NO"],
-      usesThirdPartyPersonalData: ["NO"],
-    },
-    uncompleted: {
-      purpose: ["INSTITUTIONAL"],
-      usesPersonalData: ["NO"],
-      usesThirdPartyPersonalData: ["NO"],
-    },
-  },
-  PA: {
-    version: "3.0",
-    completed: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      personalDataTypes: ["WITH_NON_IDENTIFYING_DATA"],
-      legalBasis: ["CONSENT"],
-      knowsDataQuantity: ["NO"],
-      deliveryMethod: ["CLEARTEXT"],
-      policyProvided: ["YES"],
-      confirmPricipleIntegrityAndDiscretion: ["true"],
-      doneDpia: ["NO"],
-      dataDownload: ["NO"],
-      purposePursuit: ["MERE_CORRECTNESS"],
-      checkedExistenceMereCorrectnessInteropCatalogue: ["true"],
-      usesThirdPartyData: ["NO"],
-      declarationConfirmGDPR: ["true"],
-    },
-    uncompleted: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      legalBasis: ["CONSENT"],
-      knowsDataQuantity: ["NO"],
-      deliveryMethod: ["CLEARTEXT"],
-      policyProvided: ["YES"],
-      confirmPricipleIntegrityAndDiscretion: ["true"],
-      doneDpia: ["NO"],
-      dataDownload: ["NO"],
-      purposePursuit: ["MERE_CORRECTNESS"],
-      checkedExistenceMereCorrectnessInteropCatalogue: ["true"],
-      usesThirdPartyData: ["NO"],
-      declarationConfirmGDPR: ["true"],
-    },
-  },
-};
-
-export function getRiskAnalysis({
-  tenantType,
-  completed,
-}: {
-  tenantType: TenantType;
-  completed?: boolean;
-}) {
-  const templateType =
-    tenantType === "PA1" || tenantType === "PA2" ? "PA" : "Privato/GSP";
-  const templateStatus = completed ?? true ? "completed" : "uncompleted";
-
-  const answers = RISK_ANALYSIS_DATA[templateType][templateStatus];
-  const version = RISK_ANALYSIS_DATA[templateType].version;
-
-  return {
-    name: "finalità test",
-    riskAnalysisForm: {
-      version,
-      answers,
-    },
-  };
-}
 
 export const getRandomInt = () =>
   Number(Math.random() * Number.MAX_SAFE_INTEGER).toFixed(0);
@@ -162,27 +81,4 @@ export function assertValidResponse<T>(response: AxiosResponse<T>) {
       )}`
     );
   }
-}
-
-export type FileType = "yaml" | "wsdl";
-
-export async function uploadInterfaceDocument(
-  fileName: string,
-  eserviceId: string,
-  descriptorId: string,
-  token: string
-): Promise<AxiosResponse<CreatedResource>> {
-  const blobFile = new Blob([readFileSync(`./data/${fileName}`)]);
-  const file = new File([blobFile], fileName);
-
-  return apiClient.eservices.createEServiceDocument(
-    eserviceId,
-    descriptorId,
-    {
-      kind: "INTERFACE",
-      prettyName: "Interfaccia",
-      doc: file,
-    },
-    getAuthorizationHeader(token)
-  );
 }
